Change default for `globally_enabled` to False:
This ensures we have a similar behavior to before the `globally_enabled`
flag was reworked to actually work as a global flag.

Before, even though the default was True, it only really had an effect
after saving the control panel for the first time - therefore, the actual
default was that even after installing the collective.googleauthenticator
profile 2FA was disabled.